쿠폰, 캠페인 삭제

+++ crm/src/main/java/com/encore/thecatch/event/service/EventService.java
@@ -127,7 +127,12 @@ public class EventService {
         Event event = eventRepository.findById(id).orElseThrow(
                 () -> new CatchException(ResponseCode.EVENT_NOT_FOUND)
         );
-        eventRepository.delete(event);
+        if (event.getEventStatus().equals(EventStatus.ISSUANCE)){
+            eventRepository.delete(event);
+        }else{
+            throw new CatchException(ResponseCode.EVENT_CAN_NOT_DELETE);
+        }
+
     }
 
     @PreAuthorize("hasAnyAuthority('MARKETER','ADMIN')")
